Moved form ComonJS to Typescript
