+++ .gitignore
@@ -2,3 +2,4 @@
 package-lock.json
 package.json
 /__test__
+vertical-stacked-navigation-card.js
