+++ hacs.json
@@ -1,6 +1,5 @@
 {
   "name": "Vertical Stacked Navigation Card",
-  "content_in_root": true,
   "render_readme": true,
   "filename": "vertical-stacked-navigation-card.js"
 }
