+++ src/index.ts
@@ -1,0 +1,1 @@
+export { VerticalStackedNavCard } from "./vertical-stacked-navigation-card";
