+++ src/styles.ts
@@ -1,0 +1,76 @@
+import { css } from 'lit-element';
+
+const styles = css`
+  ha-card header {
+    padding-top: 10px;
+    padding-left: 16px;
+    padding-bottom: 20px;
+    font-weight: bold;
+    font-size: 20px;
+  }
+  .nav-item-container {
+    margin-bottom: 5px;
+  }
+  .nav-item {
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    background-color: var(--main-background-color, rgba(255, 255, 255, 0.2));
+    padding: 10px;
+    color: var(--main-text-color, #fff);
+    border-radius: 5px;
+    font-size: var(--main-font-size, 20px);
+    position: relative;
+  }
+  .nav-item.unfolded {
+    border-bottom-right-radius: none !important;
+  }
+  .sub-item {
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    color: var(--sub-text-color, #fff);
+    position: relative;
+    padding: 5px;
+    background-color: var(--sub-background-color, rgba(255, 255, 255, 0.2));
+    margin-left: 15px;
+    padding-left: 25px;
+    font-size: var(--sub-font-size, 14px);
+    border-radius: none;
+  }
+  .sub-item:first-child {
+    border-top-left-radius: 3px;
+    border-top-right-radius: 3px;
+  }
+  .sub-item:last-child {
+    border-bottom-left-radius: 3px;
+    border-bottom-right-radius: 3px;
+  }
+  .nav-item:hover {
+    background-color: var(--main-hover-color, rgba(255, 255, 255, 0.3));
+  }
+  .nav-item.active {
+    background-color: var(--main-active-color, rgba(55, 55, 255, 0.5));
+  }
+  .subnav-indicator {
+    position: absolute;
+    right: 5px;
+  }
+  ha-icon {
+    margin-right: 5px;
+  }
+  .sub-nav-items {
+    display: none;
+  }
+  .sub-nav-items.open {
+    display: block;
+  }
+  .sub-item:hover {
+    background-color: var(--sub-hover-color, rgba(255, 255, 255, 0.3));
+  }
+  .sub-item.active {
+    background-color: var(--sub-active-color, rgba(55, 55, 255, 0.5));
+  }
+`;
+
+export default styles;
