+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+      "module": "ESNext",
+      "target": "ES2019",
+      "lib": ["DOM", "ES2019"],
+      "strict": true,
+      "moduleResolution": "node",
+      "esModuleInterop": true,
+      "outDir": "dist"
+    },
+    "include": ["src/**/*.ts"]
+  }
+
